Added the google auth and fist comands
When the program is run for the first time it open a browser for
autenication and then stores that autentication in a json file located
in ~/.credentials.

Now the proyect ls ten files of the google drive of the loged person
with the command ls and it quits with exit or quit.